+++ Program.cs
@@ -1,24 +1,35 @@
 ﻿using Avalonia;
+using NullWave.Helpers;
 using System;
 
 namespace NullWave;
 
 class Program
 {
-    // Initialization code. Don't use any Avalonia, third-party APIs or any
-    // SynchronizationContext-reliant code before AppMain is called: things aren't initialized
-    // yet and stuff might break.
     [STAThread]
-    public static void Main(string[] args) => BuildAvaloniaApp()
-        .StartWithClassicDesktopLifetime(args);
+    public static void Main(string[] args)
+    {
+        AppLogger.Initialize();
+        try
+        {
+            BuildAvaloniaApp().StartWithClassicDesktopLifetime(args);
+        }
+        catch (Exception ex)
+        {
+            Serilog.Log.Fatal(ex, "Application crashed");
+        }
+        finally
+        {
+            AppLogger.Shutdown();
+        }
+    }
 
-    // Avalonia configuration, don't remove; also used by visual designer.
     public static AppBuilder BuildAvaloniaApp()
         => AppBuilder.Configure<App>()
             .UsePlatformDetect()
-#if DEBUG
+            #if DEBUG
             .WithDeveloperTools()
-#endif
+            #endif
             .WithInterFont()
             .LogToTrace();
 }